Move modals controllers into separate folder
