+++ src/ggrc/assets/javascripts/controllers/modals/archive_modal_controller.js
