+++ src/ggrc/assets/javascripts/controllers/modals/delete_modal_controller.js
