+++ src/ggrc/assets/javascripts/controllers/modals/gapi-modal.js
