+++ src/ggrc/assets/javascripts/controllers/modals/help_controller.js
