+++ src/ggrc/assets/javascripts/controllers/modals/index.js
@@ -1,6 +1,6 @@
-/*!
-    Copyright (C) 2017 Google Inc.
-    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+/*
+  Copyright (C) 2017 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
 import './modals_controller';
