+++ src/ggrc/assets/javascripts/controllers/modals/quick_form_controller.js
