+++ src/ggrc/assets/javascripts/controllers/modals/unmap_modal_controller.js
