+++ src/ggrc/assets/javascripts/controllers/tests/modals_controller_spec.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import ModalsController from '../modals_controller';
+import ModalsController from '../modals/modals_controller';
 
 describe('ModalsController', function () {
   'use strict';
